added

+++ Full_Stack_CRUD_Project.js/FullstackCrud2/controllers/note.route.js
@@ -17,26 +17,34 @@ noteRouter.post("/create", async (req, res) => {
 
 
 //need to check who is going to acsess which thing ...relationship verification
-//payload can help mean decode the token again>?
+//payload can help mean decode the token again>? no man 
+// we can get using req.body.authorID bcs we manupulated that previously 
+
+
 
 
 
 noteRouter.get("/", async (req, res) => {
     try {
-        let notes = await Notemodel.find()
+        let notes = await Notemodel.find({ authorID: req.body.authorID }) //u need to get those note  where same authorid is present 
         res.status(200).send(notes)
     } catch (err) {
         res.status(400).send({ "err": err.message });
     }
-
 })
 
+
 noteRouter.patch("/update/:noteID", async (req, res) => {
-    // console.log(req.params)
     const { noteID } = req.params
+    const note = await Notemodel.findOne({ _id: noteID });
     try {
-        await Notemodel.findByIdAndUpdate({ _id: noteID }, req.body)
-        res.status(200).send({ "msg": `the note with id${noteID} has been updated`})
+        if (req.body.authorID !== note.authorID) {
+            res.status(200).send({ "msg": `you are not authorized to do this action` })
+
+        } else {
+            await Notemodel.findByIdAndUpdate({ _id: noteID }, req.body)//u can update those note where same user authorid is present 
+            res.status(200).send({ "msg": `the note with id${noteID} has been updated` })
+        }
     } catch (err) {
         res.status(400).send({ "err": err.message });
     }
@@ -45,9 +53,14 @@ noteRouter.patch("/update/:noteID", async (req, res) => {
 
 noteRouter.delete("/delete/:noteID", async (req, res) => {
     const { noteID } = req.params
+    const note = await Notemodel.findOne({ _id: noteID });
     try {
-        await Notemodel.findByIdAndDelete({ _id: noteID })
-        res.status(200).send({ "msg": `the note with id${noteID} has been Deleted`})
+        if (req.body.authorID !== note.authorID) {  //this if block is written for authorization 
+            res.status(200).send({ "msg": `you are not authorized to do this action` })
+        } else {
+            await Notemodel.findByIdAndDelete({ _id: noteID })
+            res.status(200).send({ "msg": `the note with id${noteID} has been Deleted` })
+        }
     } catch (err) {
         res.status(400).send({ "err": err.message });
     }
